Moved some files in the Sde.ConsoleGems.Consoles namespace around
