+++ ConsoleGems/Consoles/Console.cs
@@ -3,7 +3,7 @@
 // Released under the MIT license - see LICENSE.txt in the repository root.
 // </copyright>
 
-namespace Sde.ConsoleGems.Consoles.Concrete
+namespace Sde.ConsoleGems.Consoles
 {
     /// <summary>
     /// Implementation of <see cref="IConsole"/> which uses
