+++ ConsoleGems/Consoles/ConsoleColourManager.cs
@@ -3,7 +3,7 @@
 // Released under the MIT license - see LICENSE.txt in the repository root.
 // </copyright>
 
-namespace Sde.ConsoleGems.Consoles.Concrete
+namespace Sde.ConsoleGems.Consoles
 {
     /// <summary>
     /// Sets the foreground and background colours of the console window.
