+++ ConsoleGems/ServiceCollectionExtensions.cs
@@ -6,8 +6,8 @@
 namespace Sde.ConsoleGems
 {
     using System.Reflection;
-    using Microsoft.CodeAnalysis.CSharp.Syntax;
     using Microsoft.Extensions.DependencyInjection;
+    using Sde.ConsoleGems.Consoles;
     using Sde.ConsoleGems.Menus;
 
     /// <summary>
